+++ deployment/updates/scripts/add_permissions/2025_07_16_update_agent_manager_permissions.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package deployment
+ * @subpackage ursa.roles_and_permissions
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$addConfig = realpath(dirname(__FILE__)) . '/../../../permissions/service.session.ini';
+passthru("php $script $addConfig");